# perf: Improve cassette writer performance