Since cassettes writer adds a certain overhead I looked for ways to improve performance.

Here some performance benchmarks I did with the current implementation (with using [hyperfine](https://github.com/sharkdp/hyperfine) on default settings and `-i` option to ignore non-zero exit codes).

#### HTTP
The built-in test server was used with the following config:

```./test_server.sh 8081 --endpoints=success,failure,path_variable,multipart,recursive,multipart,payload,upload_file```

Command: `schemathesis run --hypothesis-max-examples=1000 --hypothesis-seed=1 --hypothesis-deadline=None http://127.0.0.1:8081/swagger.yaml` + `--store-network-log=cassette.yaml` in relevant cases

No cassettes: 11.222 s (baseline)
Current implementation with C-extension: 12.504 s (**+11.4%**)
Current implementation without C-extension: 17.058 s (**+52%**)

#### WSGI
The same test server config was used but with Flask-based implementation.
Command: `PYTHONPATH=`pwd` schemathesis run --app=test.apps._flask.app:app --hypothesis-max-examples=1000 --hypothesis-seed=1 --hypothesis-deadline=None /swagger.yaml`  + `--store-network-log=cassette.yaml` in relevant cases

No cassettes: 8.188 s (baseline)
Current implementation with C-extension: 10.616 s (**+29.6%**)
Current implementation without C-extension: 14.252 s (**+74%**)

#### Investigated approaches

1. Rust. Unfortunately `yaml_rust` [does not expose](https://github.com/chyh1990/yaml-rust/blob/master/src/emitter.rs#L159) `emit_node` and it is problematic to have incremental writing. Then `serde-yaml` doesn't expose `Serializer` which makes things even more complicated. So, I gave up there

2. Composing strings instead of using `pyyaml`. We have a simple format and it turned out that composing  YAML with string formatting is much simpler to implement than it was with the current approach.

Here are results for the string-based approach

HTTP: 11.512 s (**+2.5%**)
WSGI: 9.740 s (**+18.9%**)

I feel that this approach is a bit hacky since YAML is composed manually, but having a much simpler implementation than `pyyaml`-based one and that performance gain makes me think that it worth having.